services/tts/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
transformers==4.38.2
torch==2.2.1
soundfile==0.12.1
typing-extensions>=4.7.0
opea-comps==1.2
numpy>=1.24.0
gTTS==2.5.1 